refactor: moved getPriceRange util function out of the revisions plugin and into the catalog plugin, updated where it's being imported
